Simplify lfilter functional
* use `torch.clamp` instead of `torch.min(..., torch.max(...))`
* remove unneeded creation of ones tensor for previous method

The current lfilter function uses min and max to essentially do a clamp
function.  I changed the code to use clamp instead.  It is more readable
than the previous version.

FYI, if you want to keep the previous way, you could make a
broadcastable tensor of size 1 instead of creating a tensor the size of
the input.

Signed-off-by: David Pollack <david@da3.net>