Add SignalInfo typedef, and extension module (#718)
This is a part of PRs to add new "sox_io" backend. https://github.com/pytorch/audio/pull/726

This PR adds `SignalInfo` structure, which is data exchange interface between Python and C++ in coming TorchScript-based sox IO backend.
For the case, where C++ extension is not available (i.e. Windows), this PR also adds dummy class and module that will be substituted.
This logic is implemented in `torchaudio.extension` moduel.