Bake libsox in test base Docker image
In #728, linux unit test switches to libsox provided by apt.
For CPU jobs this is fine because all the job steps share the same Docker container,
but on CPU job, each job step runs a script in a new Docker container, so
libsox installed in a step is not available to the subsequent steps.

To fix this, this PR moves the installation of libsox and sox to Docker build.