Make `amplitude_to_DB` clamp items separately
When passed a batch, `amplitude_to_DB` was clamping based on the entire
batch's maximum value. This was wrong. Apply the clamp based on each
item's maximum when a batch is detected.

This change requires `amplitude_to_DB` to be restricted to spectrogram
inputs only, since items need to have a predictable number of
dimensions (in this case, 3) to automatically detect batches.

Additional tests are also added to check both batched and unbatched
inputs, and ensure items are being clamped correctly.