Remove walk_files (#1111)
The use of `walk_files` made it ambiguous who is responsible to locate 
the correct set of files. (Dataset class? or utility?)
In fact, just glob-ing everything is not the right problem being solved in implementing
Dataset, because if you have a specific dataset you consider to access, then
the directory structure and file locations are determined. No need to do arbitral number of recursions.
Each Dataset implementation should be glob-ing the right set of files it requires.