Simplify extension initialization
Summary:
For the case where torchaudio is used with `mode/opt`, D29973934 introduced extension initialization module specific for fbcode and the override process.

This diff simplifies the process by just inserting the step to extract extension module as a regular file at the beginning of torchaudio extension initialization process, so that OSS and fbcode use the same process / code.

Reviewed By: carolineechen

Differential Revision: D29989551

fbshipit-source-id: 9f30d0a36c220f0eb669244c9bb2da1b833d6f03