[BC-Breaking] Move fine-tune specific module out of wav2vec2 encoder
Previously, the Linear module (called `readout`) used for an ASR fine-tuning
task was placed in encoder module. Conceptually, the encoder has nothing to
do with a module specific to fine-tuning / downstreaming task.

The problems here are that;
1. encoder can be also used in pretraining phase, which does not require
such a module.
2. The choice of Linear module is arbitral, and it is inconvenient for users
to have hard-coded module structure in encoder.

Therefore this commit moves the Linear module out the encoder, and places it
as `aux` attribute of `Wav2Vec2Model`. (as a result `Wav2Vec2Model` has
`feature_extractor`, `encoder` and `aux` moduels)

An alternative approach is to define another module and place `Wav2Vec2Model`
and aux module along each other. But that will introduce a new class we need
to maintain, and since the use of `aux` is only expected for loading the
pre-trained parameters published by `fairseq` (and it's variations from HF),
and it is not general enough for downstream adoptations, where there will be
a bunch of different more complicated models. (i.e. s3prl)

So based on the minimalistic approach, we put them inside of `Wav2Vec2Model`.