Rename factory functions `wav2vec2_asr_YYY` to `wav2vec2_ft_YYY`
In #1783, we split the factory functions of wav2vec2 into ones for pretraining models
and ones for fine-tuning models (pretraining model + extra Linear module).

I picked the name scheme `wav2vec2_asr_YYY` for factory functions of fine-tuning models,
but did not feel right, because the model architecture was used for ASR fine-tuning,
but it does not have to be ASR.
This became more evident as we add pre-trained parameters support, such as #1799.
It matters more for the weight files that for which task and on which dataset it was
trained on. For factory function, ASR task is not relevant.

Therefore renaming the functions by replacing `_asr_` to `_ft_` fine-tuning.