Allow the customization of axis exclusion for ASR head
In Wav2Vec2 ASR pipelines, the `get_model` method performs on-the-fly model
surgery to remove unused dimensions common to all the Wav2Vec2 model trained
with fairseq.

In VoxPopuli, there seems to be an extra dimensions introduced due to
some issue in the preprocessing stage.

For example, the label `1` shows up in the training dataset of German (1 out of 16M),
English (1 / 28M), Spanish (1 / 9.4M), Romanian (1 / 4.7M) and Polish (6 / 5.8M).

This code changes will allow the customization of excluded dimensions for such cases.