Add randomness to test sinusoid data
This commit update `get_sinusoid` function in test utility so that
when a multi channel is requested, non-primal channel have randomized
initial phase.

This adds variety in test data. Currently `get_sinusoid` returns identical
waveforms for all the channels. The multi channel support was added just to
mock the input data so that it is easy to test features with multi-channel
inputs.

When working on numerical parity, it is more useful if the raw waveforms
are somewhat different.